refactor: extract text-field component
