--- lib/src/message-input/message-input.scss
@@ -1,149 +1,0 @@
-/* ************************
-** Variables
-** ********************* */
-:root {
-  --msg-input__background: #f8f6fd;
-  --msg-input__padding: 0 24px;
-  --msg-input__emoji-picker__bottom: 0;
-  --msg-input__emoji-picker__right: 0;
-  --msg-input__icon__color: #585858;
-  --msg-input__icon__margin: 5px 10px;
-  --msg-input__icon__fontSize: 22px;
-  --msg-input__indicator__color: #585858;
-  --msg-input__indicator__fontFamily: inherit;
-  --msg-input__indicator__fontSize: 11px;
-  --msg-input__indicator__fontWeight: normal;
-  --msg-input__indicator__margin: 10px 0 0;
-  --msg-input__indicator__minHeight: 16px;
-  --msg-input__indicator__padding: 0;
-  --msg-input__send__background: #8e2de2;
-  --msg-input__send__border: none;
-  --msg-input__send__borderRadius: 5px;
-  --msg-input__send__color: #ffffff;
-  --msg-input__send__fontFamily: inherit;
-  --msg-input__send__fontSize: 13px;
-  --msg-input__send__fontWeight: bold;
-  --msg-input__send__margin: 2px;
-  --msg-input__send__minWidth: 40px;
-  --msg-input__send__padding: 10px;
-  --msg-input__textarea--focus__border: 1px solid #9b9b9b;
-  --msg-input__textarea--focus__outline: none;
-  --msg-input__textarea__background: transparent;
-  --msg-input__textarea__border: 1px solid #9b9b9b;
-  --msg-input__textarea__borderRadius: 5px;
-  --msg-input__textarea__boxShadow: none;
-  --msg-input__textarea__color: #585858;
-  --msg-input__textarea__fontFamily: inherit;
-  --msg-input__textarea__fontSize: 13px;
-  --msg-input__textarea__lineHeight: 15px;
-  --msg-input__textarea__outline: none;
-  --msg-input__textarea__padding: 12px;
-  --msg-input__textarea__resize: none;
-  --msg-input__textarea__width: 100%;
-}
-
-/* ************************
-** Styles
-** ********************* */
-
-.pn-msg-input {
-  background: var(--msg-input__background);
-  padding: var(--msg-input__padding);
-
-  &.disabled {
-    pointer-events: none;
-  }
-
-  &__wrapper {
-    background: var(--msg-input__textarea__background);
-    border-radius: var(--msg-input__textarea__borderRadius);
-    border: var(--msg-input__textarea__border);
-    box-shadow: var(--msg-input__textarea__boxShadow);
-    display: flex;
-    position: relative;
-
-    &:focus-within {
-      outline: var(--msg-input__textarea--focus__outline);
-      border: var(--msg-input__textarea--focus__border);
-    }
-  }
-
-  &__spacer {
-    padding: var(--msg-input__textarea__padding);
-    width: 100%;
-  }
-
-  &__textarea {
-    background: transparent;
-    border: none;
-    box-sizing: border-box;
-    color: var(--msg-input__textarea__color);
-    font-family: var(--msg-input__textarea__fontFamily);
-    font-size: var(--msg-input__textarea__fontSize);
-    height: var(--msg-input__textarea__lineHeight);
-    line-height: var(--msg-input__textarea__lineHeight);
-    outline: var(--msg-input__textarea__outline);
-    padding: 0;
-    resize: var(--msg-input__textarea__resize);
-    width: var(--msg-input__textarea__width);
-  }
-
-  &__send {
-    background: var(--msg-input__send__background);
-    border-radius: var(--msg-input__send__borderRadius);
-    border: var(--msg-input__send__border);
-    color: var(--msg-input__send__color);
-    cursor: pointer;
-    flex-shrink: 0;
-    font-family: var(--msg-input__send__fontFamily);
-    font-size: var(--msg-input__send__fontSize);
-    font-weight: var(--msg-input__send__fontWeight);
-    margin: var(--msg-input__send__margin);
-    min-width: var(--msg-input__send__minWidth);
-    outline: none;
-    padding: var(--msg-input__send__padding);
-    white-space: nowrap;
-  }
-
-  &__emoji-picker {
-    bottom: var(--msg-input__emoji-picker__bottom);
-    right: var(--msg-input__emoji-picker__right);
-    position: absolute;
-  }
-
-  &__icon,
-  &__action {
-    align-self: center;
-    cursor: pointer;
-    color: var(--msg-input__icon__color);
-    font-size: var(--msg-input__icon__fontSize);
-    margin: var(--msg-input__icon__margin);
-  }
-
-  &__indicator {
-    color: var(--msg-input__indicator__color);
-    font-family: var(--msg-input__indicator__fontFamily);
-    font-size: var(--msg-input__indicator__fontSize);
-    font-weight: var(--msg-input__indicator__fontWeight);
-    margin: var(--msg-input__indicator__margin);
-    min-height: var(--msg-input__indicator__minHeight);
-    padding: var(--msg-input__indicator__padding);
-  }
-}
-
-/* ************************
-** Themes
-** ********************* */
-
-.pn-msg-input--dark,
-.pn-msg-input--event-dark,
-.pn-msg-input--support-dark {
-  --msg-input__background: #2f3136;
-  --msg-input__icon__color: #dcddde;
-  --msg-input__indicator__color: #dcddde;
-  --msg-input__textarea__background: #36393f;
-  --msg-input__textarea__border: 1px solid #36393f;
-  --msg-input__textarea__boxShadow: 0 6px 10px rgba(20, 20, 20, 0.15);
-  --msg-input__textarea__color: #dcddde;
-  --msg-input__textarea--focus__border: 1px solid #36393f;
-}
