+++ lib/src/text-field/text-field.scss
@@ -1,0 +1,150 @@
+/* ************************
+** Variables
+** ********************* */
+:root {
+    --text-field__background: #f8f6fd;
+    --text-field__padding: 0 24px;
+    --text-field__emoji-picker__bottom: 0;
+    --text-field__emoji-picker__right: 0;
+    --text-field__icon__color: #585858;
+    --text-field__icon__margin: 5px 10px;
+    --text-field__icon__fontSize: 22px;
+    --text-field__indicator__color: #585858;
+    --text-field__indicator__fontFamily: inherit;
+    --text-field__indicator__fontSize: 11px;
+    --text-field__indicator__fontWeight: normal;
+    --text-field__indicator__margin: 10px 0 0;
+    --text-field__indicator__minHeight: 16px;
+    --text-field__indicator__padding: 0;
+    --text-field__send__background: #8e2de2;
+    --text-field__send__border: none;
+    --text-field__send__borderRadius: 5px;
+    --text-field__send__color: #ffffff;
+    --text-field__send__fontFamily: inherit;
+    --text-field__send__fontSize: 13px;
+    --text-field__send__fontWeight: bold;
+    --text-field__send__margin: 2px;
+    --text-field__send__minWidth: 40px;
+    --text-field__send__padding: 10px;
+    --text-field__textarea--focus__border: 1px solid #9b9b9b;
+    --text-field__textarea--focus__outline: none;
+    --text-field__textarea__background: transparent;
+    --text-field__textarea__border: 1px solid #9b9b9b;
+    --text-field__textarea__borderRadius: 5px;
+    --text-field__textarea__boxShadow: none;
+    --text-field__textarea__color: #585858;
+    --text-field__textarea__fontFamily: inherit;
+    --text-field__textarea__fontSize: 13px;
+    --text-field__textarea__lineHeight: 15px;
+    --text-field__textarea__outline: none;
+    --text-field__textarea__padding: 12px;
+    --text-field__textarea__resize: none;
+    --text-field__textarea__width: 100%;
+  }
+  
+  /* ************************
+  ** Styles
+  ** ********************* */
+  
+  .pn-msg-input {
+    background: var(--text-field__background);
+    padding: var(--text-field__padding);
+  
+    &.disabled {
+      pointer-events: none;
+    }
+  
+    &__wrapper {
+      background: var(--text-field__textarea__background);
+      border-radius: var(--text-field__textarea__borderRadius);
+      border: var(--text-field__textarea__border);
+      box-shadow: var(--text-field__textarea__boxShadow);
+      display: flex;
+      position: relative;
+  
+      &:focus-within {
+        outline: var(--text-field__textarea--focus__outline);
+        border: var(--text-field__textarea--focus__border);
+      }
+    }
+  
+    &__spacer {
+      padding: var(--text-field__textarea__padding);
+      width: 100%;
+    }
+  
+    &__textarea {
+      background: transparent;
+      border: none;
+      box-sizing: border-box;
+      color: var(--text-field__textarea__color);
+      font-family: var(--text-field__textarea__fontFamily);
+      font-size: var(--text-field__textarea__fontSize);
+      height: var(--text-field__textarea__lineHeight);
+      line-height: var(--text-field__textarea__lineHeight);
+      outline: var(--text-field__textarea__outline);
+      padding: 0;
+      resize: var(--text-field__textarea__resize);
+      width: var(--text-field__textarea__width);
+    }
+  
+    &__send {
+      background: var(--text-field__send__background);
+      border-radius: var(--text-field__send__borderRadius);
+      border: var(--text-field__send__border);
+      color: var(--text-field__send__color);
+      cursor: pointer;
+      flex-shrink: 0;
+      font-family: var(--text-field__send__fontFamily);
+      font-size: var(--text-field__send__fontSize);
+      font-weight: var(--text-field__send__fontWeight);
+      margin: var(--text-field__send__margin);
+      min-width: var(--text-field__send__minWidth);
+      outline: none;
+      padding: var(--text-field__send__padding);
+      white-space: nowrap;
+    }
+  
+    &__emoji-picker {
+      bottom: var(--text-field__emoji-picker__bottom);
+      right: var(--text-field__emoji-picker__right);
+      position: absolute;
+    }
+  
+    &__icon,
+    &__action {
+      align-self: center;
+      cursor: pointer;
+      color: var(--text-field__icon__color);
+      font-size: var(--text-field__icon__fontSize);
+      margin: var(--text-field__icon__margin);
+    }
+  
+    &__indicator {
+      color: var(--text-field__indicator__color);
+      font-family: var(--text-field__indicator__fontFamily);
+      font-size: var(--text-field__indicator__fontSize);
+      font-weight: var(--text-field__indicator__fontWeight);
+      margin: var(--text-field__indicator__margin);
+      min-height: var(--text-field__indicator__minHeight);
+      padding: var(--text-field__indicator__padding);
+    }
+  }
+  
+  /* ************************
+  ** Themes
+  ** ********************* */
+  
+  .pn-msg-input--dark,
+  .pn-msg-input--event-dark,
+  .pn-msg-input--support-dark {
+    --text-field__background: #2f3136;
+    --text-field__icon__color: #dcddde;
+    --text-field__indicator__color: #dcddde;
+    --text-field__textarea__background: #36393f;
+    --text-field__textarea__border: 1px solid #36393f;
+    --text-field__textarea__boxShadow: 0 6px 10px rgba(20, 20, 20, 0.15);
+    --text-field__textarea__color: #dcddde;
+    --text-field__textarea--focus__border: 1px solid #36393f;
+  }
+
