+++ samples/src/event-chat/event-chat.scss
@@ -88,6 +88,6 @@
   }
 
   .pn-msg-input {
-    --msg-input__padding: 0 10px 15px;
+    --text-field__padding: 0 10px 15px;
   }
 }
