+++ samples/src/group-chat/group-chat.scss
@@ -55,6 +55,6 @@
   }
 
   .pn-msg-input {
-    --msg-input__padding: 0 24px 24px;
+    --text-field__padding: 0 24px 24px;
   }
 }
