+++ samples/src/simple-chat/simple-chat.scss
@@ -9,8 +9,8 @@
   --app-people-active-color: #5babfc;
 
   .pn-msg-input {
-    --msg-input__background: transparent;
-    --msg-input__send__background: #5babfc;
+    --text-field__background: transparent;
+    --text-field__send__background: #5babfc;
   }
 
   .pn-msg-list--light {
@@ -38,7 +38,7 @@
   --app-people-active-color: #de2440;
 
   .pn-msg-input {
-    --msg-input__send__background: #de2440;
+    --text-field__send__background: #de2440;
   }
 
   .pn-msg-list--dark {
@@ -62,10 +62,10 @@
   overflow: hidden;
 
   .pn-msg-input {
-    --msg-input__send__borderRadius: 25px;
-    --msg-input__send__padding: 10px 20px;
-    --msg-input__textarea__borderRadius: 25px;
-    --msg-input__textarea__padding: 12px 24px;
+    --text-field__send__borderRadius: 25px;
+    --text-field__send__padding: 10px 20px;
+    --text-field__textarea__borderRadius: 25px;
+    --text-field__textarea__padding: 12px 24px;
   }
 
   .pn-channel-list {
